Make universal default prompt plain
fixed #15006
